requirements.txt
amqp==5.1.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi @ file:///croot/certifi_1665076670883/work/certifi
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Flask==2.2.2
greenlet==1.1.2
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
prompt-toolkit==3.0.36
pytz==2022.6
redis==4.4.0
redis-py-cluster==2.1.2
six==1.16.0
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
mongoengine==0.22.1
python-dateutil